Move EncryptionException to core
to avoid the ClassNotFound failure when the exception is forwarded from daemon to client
